refactor(text-to-image): extracts `firstTextToImageOutput({ in, inferredFrom })`
